+++ js/tree.js
@@ -9,7 +9,7 @@ var layoutInfo = {
 
 
 // A "ghost" layer which offsets other layers in the tree
-addNode("spook", {
+addNode("blank", {
     layerShown: "ghost",
 }, 
 )